<ADD> 6단계 테스트 코드 추가 및 테스트 깨지는 부분 수정
