--- src/test/java/roomescape/SpringDataJPATest.java
@@ -1,89 +1,0 @@
-package roomescape;
-
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import roomescape.reservation.MyReservationResponse;
-import roomescape.time.Time;
-import roomescape.time.TimeRepository;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@DataJpaTest
-public class SpringDataJPATest {
-
-    @Autowired
-    private TestEntityManager entityManager;
-
-    @Autowired
-    private TimeRepository timeRepository;
-
-    @Test
-    void 사단계() {
-        Time time = new Time("10:00");
-        entityManager.persist(time);
-        entityManager.flush();
-
-        Time persistTime = timeRepository.findById(time.getId()).orElse(null);
-
-        assertThat(persistTime.getValue()).isEqualTo(time.getValue());
-    }
-
-    @Test
-    void 오단계() {
-        RestAssured.port = 8080;
-
-        String adminToken = createToken("admin@email.com", "password");
-
-        List<MyReservationResponse> reservations = RestAssured.given().log().all()
-                .cookie("token", adminToken)
-                .get("/reservations-mine")
-                .then().log().all()
-                .statusCode(200)
-                .extract().jsonPath().getList(".", MyReservationResponse.class);
-
-        assertThat(reservations).hasSize(3);
-    }
-
-    @Test
-    void 오단계_추가테스트() {
-        RestAssured.port = 8080;
-
-        String adminToken = createToken("brown@email.com", "password");
-
-        List<MyReservationResponse> reservations = RestAssured.given().log().all()
-                .cookie("token", adminToken)
-                .get("/reservations-mine")
-                .then().log().all()
-                .statusCode(200)
-                .extract().jsonPath().getList(".", MyReservationResponse.class);
-
-        assertThat(reservations).hasSize(1);
-    }
-
-    private String createToken(String email, String password) {
-        Map<String, String> params = new HashMap<>();
-        params.put("email", email);
-        params.put("password", password);
-
-        ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/login")
-                .then().log().all()
-                .statusCode(200)
-                .extract();
-
-        return response.headers().get("Set-Cookie").getValue().split(";")[0].split("=")[1];
-    }
-}
-
